requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bottleneck=1.3.4=py39h67323c0_0
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.4.26=hecd8cb5_0
certifi=2022.5.18.1=py39hecd8cb5_0
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
ffmpeg=4.3=h0a44026_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
kiwisolver=1.4.2=py39he9d5cce_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=hdb42f99_1
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
matplotlib=3.5.1=py39hecd8cb5_1
matplotlib-base=3.5.1=py39hfb0c5b7_1
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_2
nettle=3.7.3=h230ac6f_1
numexpr=2.8.1=py39h2e5f0a9_0
numpy=1.22.3=py39h2e5f0a9_0
numpy-base=1.22.3=py39h3b1a694_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1o=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39he9d5cce_0
pillow=9.0.1=py39hde71d04_0
pip=21.2.4=py39hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.12=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.11.0=py3.9_0
pytz=2021.3=pyhd3eb1b0_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.7.3=py39h8c7af03_0
setuptools=61.2.0=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=h707629a_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h3fd3227_1
torchaudio=0.11.0=py39_cpu
torchvision=0.12.0=py39_cpu
tornado=6.1=py39h9ed2024_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39hecd8cb5_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=hca72f7f_1
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
